requirements.txt
streamlit==1.39.0
pypdf==5.1.0
pandas==2.2.3
requests==2.32.3
duckduckgo-search==6.3.5
python-dotenv==1.0.1
openai==1.52.2
